Rename Api endpoints
