+++ controllers/api/index.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
-const userRoutes = require('./user');
-const categoriesRoutes = require('./categories');
-const transactionRoutes = require('./transaction');
+const userRoutes = require('./userRoutes');
+const categoriesRoutes = require('./categoryRoutes');
+const transactionRoutes = require('./transactionRoutes');
 
 router.use('/user', userRoutes);
 router.use('/categories', categoriesRoutes);
